+++ tf2onnx/custom_opsets/__init__.py
@@ -4,7 +4,3 @@
 
 from . import ms
 from .. import constants
-
-DOMAIN_OPSETS = {
-    constants.MICROSOFT_DOMAIN: ms.OPSETS
-}
