--- tf2onnx/function/gathernd.py
@@ -1,153 +1,0 @@
-# Copyright (c) Microsoft Corporation. All rights reserved.
-# Licensed under the MIT license.
-
-"""
-tf2onnx.tf2onnx - gathernd op conversion
-"""
-import numpy as np
-from onnx.onnx_pb import TensorProto
-from tf2onnx import utils
-
-# pylint: disable=unused-argument,missing-docstring
-
-INT64_MAX = np.iinfo(np.int64).max
-
-
-def _make_gathernd_inner_loop(ctx, params, index, dtype):
-    """create the inner loop for GatherNd."""
-    # gather_cur = params
-    # for (int i = 0; i < size(index); i++)
-    #   gather_res = gather(gather_cur, index[i])
-    scope_name = utils.make_name("gathernd_inner_loop")
-    trip_node = ctx.make_node("Size", [index.output[0]])
-    cond_const = ctx.make_const(utils.make_name("cond"), np.ones((), dtype=np.bool))
-    trip_name = utils.make_name("i")
-    cond_name = utils.make_name("cond")
-    cond_out_name = utils.make_name("cond_out")
-    cur_name = utils.make_name("gather_cur")
-    result_name = utils.make_name("res")
-
-    # body graph creation
-    g = ctx.create_new_graph_with_same_config()
-    index_i = g.make_node("Gather", [index.output[0], trip_name], attr={"axis": 0})
-    gather = g.make_node("Gather", [cur_name, index_i.output[0]], attr={"axis": 0})
-    g.make_node("Squeeze", [gather.output[0]], attr={"axes": [0]}, outputs=[result_name])
-    g.make_node("Identity", [cond_name], outputs=[cond_out_name])
-
-    g.add_graph_input(trip_name, TensorProto.INT64, [])
-    g.add_graph_input(cond_name, TensorProto.BOOL, [])
-    g.add_graph_input(cur_name, dtype, [])
-
-    g.add_graph_output(cond_out_name, TensorProto.BOOL, [])
-    g.add_graph_output(result_name, dtype, [])
-
-    inner_loop = ctx.make_node("Loop", [trip_node.output[0],
-                                        cond_const.output[0],
-                                        params],
-                               op_name_scope=scope_name, skip_conversion=False)
-    inner_loop.set_body_graph_as_attr("body", g)
-    return inner_loop
-
-
-def make_gathernd(ctx, params, indices, output, scope_name, t_params, shapes, dtypes):
-    """make GatherNd op."""
-    # Tparams output = GatherNd(Tparams params, Tidx indices)
-    scope_name = utils.make_name(scope_name)
-    # reshape indices into [sum(indices[:-1]), indices[-1]]
-    indices_shape = ctx.make_node("Shape", [indices], dtypes=[TensorProto.INT64])
-    indices_size = ctx.make_node("Size", [indices])
-    inner_shape = ctx.make_node("Slice",
-                                [indices_shape.output[0]],
-                                attr={"axes": [0], "ends": [INT64_MAX], "starts": [-1]},
-                                dtypes=[TensorProto.INT64])
-    outter_shape = ctx.make_node("Div",
-                                 [indices_size.output[0], inner_shape.output[0]],
-                                 dtypes=[TensorProto.INT64])
-    flatten_shape = ctx.make_node("Concat",
-                                  [outter_shape.output[0], inner_shape.output[0]],
-                                  attr={"axis": 0},
-                                  dtypes=[TensorProto.INT64])
-    flatten_indices = ctx.make_node("Reshape", [indices, flatten_shape.output[0]])
-
-    # outter loop for each index
-    # for (int i=0; i<outter_shape; i++) inner_loop(params, flatten_indices[i])
-    cond_const = ctx.make_const(utils.make_name("cond"), np.ones((), dtype=np.bool))
-    dummy_const = ctx.make_const(utils.make_name("dummy"), np.ones((), dtype=np.int64))
-
-    # body graph creation
-    g = ctx.create_new_graph_with_same_config()
-    trip_name = utils.make_name("i")
-    cond_name = utils.make_name("cond")
-    cond_out_name = utils.make_name("cond_out")
-    dummy_name = utils.make_name("dummy")
-    dummy_out_name = utils.make_name("dummy_out")
-    result_name = utils.make_name("res")
-
-    index = g.make_node("Gather", [flatten_indices.output[0], trip_name], attr={"axis": 0})
-    index_squeeze = g.make_node("Squeeze", [index.output[0]], attr={"axes": [0]})
-    # inner loop to gather result
-    inner_loop = _make_gathernd_inner_loop(g, params, index_squeeze, t_params)
-    g.make_node("Identity", [cond_name], outputs=[cond_out_name])
-    g.make_node("Identity", [dummy_name], outputs=[dummy_out_name])
-    g.make_node("Identity", [inner_loop.output[0]], outputs=[result_name])
-
-    g.add_graph_input(trip_name, TensorProto.INT64, [])
-    g.add_graph_input(cond_name, TensorProto.BOOL, [])
-    g.add_graph_input(dummy_name, t_params, [])
-
-    g.add_graph_output(cond_out_name, TensorProto.BOOL, [])
-    g.add_graph_output(dummy_out_name, t_params, [])
-    g.add_graph_output(result_name, t_params, [])
-
-    gathernd_loop = ctx.make_node("Loop",
-                                  [outter_shape.output[0], cond_const.output[0], params],
-                                  output_count=2,
-                                  op_name_scope=scope_name, skip_conversion=False)
-    gathernd_loop.set_body_graph_as_attr("body", g)
-
-    # reshape to target shape
-    # output shape of gathernd: indices.shape[:-1] + gathernd_output.shape[1:]
-    inner_loop_shape = ctx.make_node("Shape", [gathernd_loop.output[1]], dtypes=[TensorProto.INT64])
-    # workaround in case gathernd_loop is 1-dimensional
-    one_const = ctx.make_const(utils.make_name("one"), np.array([1], dtype=np.int64))
-    inner_loop_shape_ = ctx.make_node("Concat",
-                                      [inner_loop_shape.output[0], one_const.output[0]],
-                                      attr={"axis": 0},
-                                      dtypes=[TensorProto.INT64])
-    output_inner_shape = ctx.make_node("Slice",
-                                       [inner_loop_shape_.output[0]],
-                                       attr={"axes": [0], "ends": [INT64_MAX], "starts": [1]},
-                                       dtypes=[TensorProto.INT64])
-
-    indices_outter_shape = ctx.make_node("Slice",
-                                         [indices_shape.output[0]],
-                                         attr={"axes": [0], "ends": [-1], "starts": [0]},
-                                         dtypes=[TensorProto.INT64])
-    output_shape_ = ctx.make_node("Concat",
-                                  [indices_outter_shape.output[0], output_inner_shape.output[0]],
-                                  attr={"axis": 0},
-                                  dtypes=[TensorProto.INT64])
-    output_shape = ctx.make_node("Slice",
-                                 [output_shape_.output[0]],
-                                 attr={"axes": [0], "ends": [-1], "starts": [0]},
-                                 dtypes=[TensorProto.INT64])
-    ctx.make_node("Reshape",
-                  [gathernd_loop.output[1], output_shape.output[0]],
-                  outputs=[output],
-                  shapes=shapes,
-                  dtypes=dtypes)
-
-
-def gathernd_op(ctx, node, name, args):
-    """GatherNd op."""
-    # Tparams output = GatherNd(Tparams params, Tidx indices)
-    params = node.input[0]
-    indices = node.input[1]
-    output = node.output[0]
-    # same as the attr Tparams
-    t_params = ctx.get_dtype(params)
-    utils.make_sure(t_params, "Dtype of {} is None".format(indices))
-    shapes = node.output_shapes
-    dtypes = node.output_dtypes
-    ctx.remove_node(node.name)
-    make_gathernd(ctx, params, indices, output, name, t_params, shapes, dtypes)
