+++ tf2onnx/onnx_opset/__init__.py
@@ -1,0 +1,5 @@
+# Copyright (c) Microsoft Corporation. All rights reserved.
+# Licensed under the MIT license.
+"""tf2onnx.onnx_opset module"""
+
+from . import common, controlflow, generator, logical, math, misc, nn, reduction, rnn, tensor, traditionalml
