+++ tf2onnx/onnx_opset/traditionalml.py
@@ -1,0 +1,17 @@
+# Copyright (c) Microsoft Corporation. All rights reserved.
+# Licensed under the MIT license.
+
+"""
+tf2onnx.tf2onnx.onnx_opset.generator
+"""
+
+from __future__ import division
+from __future__ import print_function
+from __future__ import unicode_literals
+
+import logging
+
+logging.basicConfig(level=logging.INFO)
+log = logging.getLogger("onnx_opset.traditional")
+
+# pylint: disable=unused-argument,missing-docstring
